refactor: 기존에 ui 패키지에 어울리지 않는 파일 위치 변경
