+++ src/main/java/roomescape/login/interceptor/AdminInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape.login.ui;
+package roomescape.login.interceptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
